refactor(knowledge): extract transfer logic into KnowledgeTransferService and add StructuredValidationException
- Extract document transfer methods from KnowledgeService into dedicated
  KnowledgeTransferService class (knowledge_transfer.py)
- Add StructuredValidationException for stable frontend-localizable error codes
- Simplify transfer namespace validation with declarative allowed-target map
- Refactor frontend transfer dialog to use TRANSFER_ALLOWED_TARGET_SCOPES map
- Streamline test suite by removing redundant verbose assertions while
  preserving core coverage
- Remove unused imports and dead code from knowledge_service.py
